requirements.txt
asgiref==3.7.2
Django==5.0.2
django-dotenv==1.4.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
